+++ source/round.cpp
@@ -1,5 +1,4 @@
 #include "../include/round.h"
-#include "../include/matchmaking.h"
 #include "../include/console_utils.h"
 
 #include <iostream>
@@ -14,10 +13,8 @@ void Round::startRound(std::vector<Player>& players) {
     std::cout << std::endl;
 
     // Matchmaking
-    std::cout << "Match making in progress..." << std::endl;
-    Matchmaker matchmaker{};
     matchmaker.makeMatches(players);
-    std::cout << "Number of matches made: " << matchmaker.matches.size() << std::endl;
+    match_status.resize(matchmaker.matches.size(), false);
 
     do {
         std::cout << std::endl;
@@ -29,31 +26,54 @@ void Round::startRound(std::vector<Player>& players) {
         // Setting results
         std::cout << std::endl;
         std::cout << "Enter match number to update its status: ";
-        unsigned int choice;
+        int choice;
         std::cin >> choice;
         std::cin.ignore(); // Flush new line
 
+        if (choice < 1 || choice > matchmaker.matches.size()) {
+            Console::setColour(12);
+            std::cout << "ERROR: Invalid input " << choice << std::endl;
+            Console::setColour(7);
+            continue;
+        }
+
         std::cout << std::endl;
         std::cout << "Match " << choice << ": " << std::endl;
-        displayResultsMenu(matchmaker.matches[choice - 1]);
-        unsigned int result;
+        displayResultsMenu(choice - 1);
+        int result;
         std::cin >> result;
         std::cin.ignore();
 
-        switch (result) {
-            case 1:
-                updateMatchState(matchmaker.matches[choice - 1][0], matchmaker.matches[choice - 1][1]);
-                break;
-            case 2:
-                updateMatchState(matchmaker.matches[choice - 1][1], matchmaker.matches[choice - 1][0]);
-                break;
-            case 3:
-                updateMatchState(matchmaker.matches[choice - 1][0], matchmaker.matches[choice - 1][1], true);
-                break;
-            default:
-                Console::setColour(12);
-                std::cout << "ERROR: Invalid input " << result << std::endl;
-                Console::setColour(7);
+        if (match_status[choice - 1]) {
+            switch (result) {
+                case 1:
+                    updateMatchState(matchmaker.matches[choice - 1][0], matchmaker.matches[choice - 1][1], choice - 1);
+                    match_status[choice - 1] = false;
+                    break;
+                default:
+                    Console::setColour(12);
+                    std::cout << "ERROR: Invalid input " << result << std::endl;
+                    Console::setColour(7);
+            }
+        } else {
+            switch (result) {
+                case 1:
+                    updateMatchState(matchmaker.matches[choice - 1][0], matchmaker.matches[choice - 1][1], choice - 1);
+                    match_status[choice - 1] = true;
+                    break;
+                case 2:
+                    updateMatchState(matchmaker.matches[choice - 1][1], matchmaker.matches[choice - 1][0], choice - 1);
+                    match_status[choice - 1] = true;
+                    break;
+                case 3:
+                    updateMatchState(matchmaker.matches[choice - 1][0], matchmaker.matches[choice - 1][1], choice - 1, true);
+                    match_status[choice - 1] = true;
+                    break;
+                default:
+                    Console::setColour(12);
+                    std::cout << "ERROR: Invalid input " << result << std::endl;
+                    Console::setColour(7);
+            }
         }
     } while (!roundOver(matchmaker.matches));
 
@@ -81,25 +101,36 @@ void Round::startRound(std::vector<Player>& players) {
     std::cout << "Round " << round << " is over!";
 }
 
-void Round::updateMatchState(Player& winner, Player& loser, bool draw) {
+void Round::updateMatchState(Player& winner, Player& loser, const int& index, bool draw) {
     std::string winner_code, loser_code;
-    if (draw) {
-        winner_code = "=" + std::to_string(loser.id);
-        loser_code = "=" + std::to_string(winner.id);
+    if (match_status[index]) {
+        winner.roundDetails.pop_back();
+        loser.roundDetails.pop_back();
     } else {
-        winner_code = "+" + std::to_string(loser.id);
-        loser_code = "-" + std::to_string(winner.id);
+        if (draw) {
+            winner_code = "=" + std::to_string(loser.id);
+            loser_code = "=" + std::to_string(winner.id);
+        } else {
+            winner_code = "+" + std::to_string(loser.id);
+            loser_code = "-" + std::to_string(winner.id);
+        }
+        winner.roundDetails.push_back(winner_code);
+        loser.roundDetails.push_back(loser_code);
     }
-    winner.roundDetails.push_back(winner_code);
-    loser.roundDetails.push_back(loser_code);
 }
 
-void Round::displayResultsMenu(const std::vector<Player>& match) {
-    std::cout << "[1] " << match[0].name << " wins" << std::endl;
-    std::cout << "[2] " << match[1].name << " wins" << std::endl;
-    std::cout << "[3] Draw" << std::endl;
-
-    std::cout << "What is the match result: ";
+void Round::displayResultsMenu(const int& index) {
+    const std::vector<Player>& match = matchmaker.matches[index];
+    if (!match_status[index]) {
+        std::cout << "[1] " << match[0].name << " wins" << std::endl;
+        std::cout << "[2] " << match[1].name << " wins" << std::endl;
+        std::cout << "[3] Draw" << std::endl;
+        std::cout << "What is the match result: ";
+    } else {
+        std::cout << "[1] Reset" << std::endl;
+        
+        std::cout << "Enter a choice: ";
+    }
 }
 
 bool Round::roundOver(const std::vector<std::vector<Player>>& matches) {