+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be noted here.
 
 ---
 
+## [1.1.1] - 24 July 2025
+
+### Bugs Fixed
+Fixed a bug where:
+- Match status could be set multiple times in a single round
+- Match number input could recieve numbers that are not valid (Non-existing table numbers)
+
+## [1.1.0] - 24 July 2025
+
+### Modified
+- List of matches is now displayed in a tabular format
+- List of matches displays players' colour
+- Final standings shows Sonneborn-Berger scores
+- Exporting and importing data considers Sonneborn-Berger scores
+
+### Added
+- Colour assignment after match making
+- Sonneborn-Berger scoring system
+
 ## [1.0.0] - 13 Jul 2025
 Introducing the first version of this project...